#21 Split Bintray class
