+++ bintray/bintray.py
@@ -1,13 +1,13 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """ Python Wrapper for Bintray API
 
     https://bintray.com/docs/api
 """
 import os
-import logging
-import requests
-from requests.auth import HTTPBasicAuth
+
+from bintray.requester import Requester
+from bintray.logger import Logger
+from bintray.utils import bool_to_number
 
 
 __version__ = "0.1.1"
@@ -31,51 +31,8 @@ class Bintray(object):
         """
         self._username = username or os.getenv("BINTRAY_USERNAME")
         self._password = api_key or os.getenv("BINTRAY_API_KEY")
-
-        self._logger = logging.getLogger(__file__)
-        self._logger.setLevel(logging.INFO)
-        formatter = logging.Formatter('%(asctime)s:%(levelname)s: %(message)s')
-        ch = logging.StreamHandler()
-        level = int(os.getenv("BINTRAY_LOGGING_LEVEL", logging.INFO))
-        ch.setLevel(level)
-        ch.setFormatter(formatter)
-        self._logger.addHandler(ch)
-
-    def _get_authentication(self):
-        """ Retrieve Basic HTTP Authentication based on username and API key
-
-        :return: Basic Authentication handler
-        """
-        if not self._username or not self._password:
-            return None
-        return HTTPBasicAuth(self._username, self._password)
-
-    def _add_status_code(self, response):
-        """ Update JSON result with error and status code
-
-        :param response: Requests response
-        :return: Response JSON
-        """
-        json_data = response.json()
-        if isinstance(json_data, list):
-            json_data.append({"statusCode": response.status_code, "error": not response.ok})
-        else:
-            json_data.update({"statusCode": response.status_code, "error": not response.ok})
-        return json_data
-
-    def _bool_to_number(self, value):
-        """ Convert boolean result into numeric string
-
-        :param value: Any boolean value
-        :return: "1" when True. Otherwise, "0"
-        """
-        return "1" if value else "0"
-
-    def _raise_error(self, message, response):
-        try:
-            response.raise_for_status()
-        except Exception as error:
-            raise Exception("{} ({}): {}".format(message, response.status_code, str(error)))
+        self._requester = Requester(self._username, self._password)
+        self._logger = Logger().logger
 
     # Files
 
@@ -90,16 +47,13 @@ class Bintray(object):
         :param include_unpublished: Show not published files
         :return: List with all files
         """
-        parameters = {"include_unpublished": self._bool_to_number(include_unpublished)}
+        parameters = {"include_unpublished": bool_to_number(include_unpublished)}
         url = "{}/packages/{}/{}/{}/files?include_unpublished={}".format(Bintray.BINTRAY_URL,
                                                                 subject,
                                                                 repo,
                                                                 package,
                                                                 include_unpublished)
-        response = requests.get(url, auth=self._get_authentication(), params=parameters)
-        if not response.ok:
-            self._raise_error("Could not list package files", response)
-        return self._add_status_code(response)
+        return self._requester.get(url, parameters)
 
     # Content Uploading & Publishing
 
@@ -121,17 +75,15 @@ class Bintray(object):
         """
         url = "{}/content/{}/{}/{}/{}/{}".format(Bintray.BINTRAY_URL, subject, repo, package,
                                                  version, remote_file_path)
-        parameters = {"publish": self._bool_to_number(publish),
-                      "override": self._bool_to_number(override),
-                      "explode": self._bool_to_number(explode)}
+        parameters = {"publish": bool_to_number(publish),
+                      "override": bool_to_number(override),
+                      "explode": bool_to_number(explode)}
 
         with open(local_file_path, 'rb') as file_content:
-            response = requests.put(url, auth=self._get_authentication(), params=parameters,
-                                    data=file_content)
-            if response.status_code != 201:
-                self._raise_error("Could not upload", response)
+            response = self._requester.put(url, params=parameters, data=file_content)
+
         self._logger.info("Upload successfully: {}".format(url))
-        return self._add_status_code(response)
+        return response
 
     # Content Downloading
 
@@ -145,13 +97,13 @@ class Bintray(object):
         """
         download_base_url = "https://dl.bintray.com"
         url = "{}/{}/{}/{}".format(download_base_url, subject, repo, remote_file_path)
-        response = requests.get(url, auth=self._get_authentication())
-        if not response.ok:
-            self._raise_error("Could not download file content", response)
+        response, content = self._requester.download(url)
+
         with open(local_file_path, 'wb') as local_fd:
-            local_fd.write(response.content)
+            local_fd.write(content)
+
         self._logger.info("Download successfully: {}".format(url))
-        return self._add_status_code(response)
+        return response
 
     # Licenses
 
@@ -161,8 +113,4 @@ class Bintray(object):
         :return: List with OSS licenses
         """
         url = "{}/licenses/oss_licenses".format(Bintray.BINTRAY_URL)
-        response = requests.get(url, auth=self._get_authentication())
-        if not response.ok:
-            self._raise_error("Could not get OSS licenses", response)
-        return self._add_status_code(response)
-
+        return self._requester.get(url)
