+++ bintray/utils.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+
+def bool_to_number(value):
+    """ Convert boolean result into numeric string
+
+    :param value: Any boolean value
+    :return: "1" when True. Otherwise, "0"
+    """
+    return "1" if value else "0"
